Create Potion: Healing, recovery mana and Sub Potion: Reduce Stamina usage,Reduce Damage taken, Increase Damage. Create UI for Item Using, and Status UI

+++ Assets/Script/Attack/AttackData.cs
@@ -13,6 +13,6 @@ public class AttackData
     [field: SerializeField] public float AnimationTransition { get; private set; }
     [field: SerializeField] public float ForceTime { get; private set; }
     [field: SerializeField] public float Force { get; private set; }
-    [field: SerializeField] public int AttackDamage { get; set; }
+    [field: SerializeField] public float AttackDamage { get; set; }
     [field: SerializeField] public SkillType SkillType { get; private set; }
 }
